fixing Member Crud

+++ src/main/java/services/MemberServiceImpl.java
@@ -1,16 +1,17 @@
 package services;
 
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
 import dao.MemberDAO;
 import dto.MemberDTO;
 import models.Member;
 import repositories.MemberRepository;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class MemberServiceImpl implements MemberService{
-    private MemberDAO memberDAO;
-    private MemberRepository memberRepository;
+public class MemberServiceImpl implements MemberService {
+    private final MemberDAO memberDAO;
+    private final MemberRepository memberRepository;
 
     public MemberServiceImpl(MemberDAO memberDAO, MemberRepository memberRepository) {
         this.memberDAO = memberDAO;
@@ -18,45 +19,38 @@ public class MemberServiceImpl implements MemberService{
     }
 
     @Override
-    public void addMember(MemberDTO memberDTO) {
+    public Member addMember(MemberDTO memberDTO) throws SQLException {
         Member member = new Member(memberDTO.getId(), memberDTO.getFname(), memberDTO.getLname(),
                 memberDTO.getEmail(), memberDTO.getRole(), memberDTO.getTeamId());
-        memberDAO.addMember(member);
-    }
-    
-    @Override
-    public List<MemberDTO> getAllMembers() {
-        List<Member> members = memberDAO.getAllMembers();
-        List<MemberDTO> memberDTOs = new ArrayList<>();
-        for (Member member : members) {
-            memberDTOs.add(new MemberDTO(member.getId(), member.getFname(), member.getLname(),
-                    member.getEmail(), member.getRole(), member.getTeamId()));
-        }
-        return memberDTOs;
-    }
-    
-    @Override
-    public List<MemberDTO> searchMembersByName(String name) {
-        List<Member> members = memberRepository.searchMembersByName(name);
-        List<MemberDTO> memberDTOs = new ArrayList<>();
-        for (Member member : members) {
-            memberDTOs.add(new MemberDTO(member.getId(), member.getFname(), member.getLname(),
-                    member.getEmail(), member.getRole(), member.getTeamId()));
-        }
-        return memberDTOs;
-    }
-    
-    @Override
-    public void updateMember(MemberDTO memberDTO) {
-        Member member = new Member(memberDTO.getId(), memberDTO.getFname(), memberDTO.getLname(),
-                memberDTO.getEmail(), memberDTO.getRole(), memberDTO.getTeamId());
-        memberDAO.updateMember(member);
+        return memberDAO.addMember(member);
     }
 
     @Override
-    public void deleteMember(int id) {
+    public Member updateMember(MemberDTO memberDTO) throws SQLException {
+        if (memberDAO.getMemberById(memberDTO.getId()) == null) {
+            return null;
+        } else {
+            Member member = new Member(memberDTO.getId(), memberDTO.getFname(), memberDTO.getLname(),
+                    memberDTO.getEmail(), memberDTO.getRole(), memberDTO.getTeamId());
+            return memberDAO.updateMember(member);
+        }
+    }
+
+    @Override
+    public void deleteMember(int id) throws SQLException {
+        if (memberDAO.getMemberById(id) == null) {
+            throw new SQLException("Member not found with ID: " + id);
+        }
         memberDAO.deleteMember(id);
     }
 
+    @Override
+    public List<Member> getAllMembers() throws SQLException {
+        return memberDAO.getAllMembers();
+    }
 
+    @Override
+    public List<Member> searchMembers(String name) throws SQLException {
+        return memberRepository.searchMembers(name);
+    }
 }
